Fix empty array encoding
The Redis protocol specification[1] defines an empty array as the
sequence `*0\r\n`. The previous value, `*-1\r\n` is what Redis calls the
"Null Array" which is used in some special cases, for example when a
`BLPOP` times out.

The variable is also renamed for consistency with the Redis conventions.

[1] https://redis.io/topics/protocol#resp-arrays